Rate Limiting refactored
