+++ obp-api/src/main/scala/code/api/v3_1_0/APIMethods310.scala
@@ -666,7 +666,7 @@ trait APIMethods310 {
             (Full(u), callContext) <-  authorizedAccess(cc)
             _ <- NewStyle.function.hasEntitlement("", u.userId, canReadCallLimits, callContext)
             consumer <- NewStyle.function.getConsumerByConsumerId(consumerId, callContext)
-            rateLimit <- Future(RateLimitUtil.consumerRateLimitState(consumer.key.get).toList)
+            rateLimit <- Future(RateLimitingUtil.consumerRateLimitState(consumer.key.get).toList)
           } yield {
             (createCallLimitJson(consumer, rateLimit), HttpCode.`200`(callContext))
           }
@@ -1297,14 +1297,14 @@ trait APIMethods310 {
           for {
             (_, callContext) <- anonymousAccess(cc)
             rateLimiting <- NewStyle.function.tryons("", 400, callContext) {
-              RateLimitUtil.inMemoryMode match {
+              RateLimitingUtil.inMemoryMode match {
                 case true =>
-                  val isActive = if(RateLimitUtil.useConsumerLimits == true) true else false
-                  RateLimiting(RateLimitUtil.useConsumerLimits, "In-Memory", true, isActive)
+                  val isActive = if(RateLimitingUtil.useConsumerLimits == true) true else false
+                  RateLimiting(RateLimitingUtil.useConsumerLimits, "In-Memory", true, isActive)
                 case false =>
-                  val isRedisAvailable = RateLimitUtil.isRedisAvailable()
-                  val isActive = if(RateLimitUtil.useConsumerLimits == true && isRedisAvailable == true) true else false
-                  RateLimiting(RateLimitUtil.useConsumerLimits, "REDIS", isRedisAvailable, isActive)
+                  val isRedisAvailable = RateLimitingUtil.isRedisAvailable()
+                  val isActive = if(RateLimitingUtil.useConsumerLimits == true && isRedisAvailable == true) true else false
+                  RateLimiting(RateLimitingUtil.useConsumerLimits, "REDIS", isRedisAvailable, isActive)
               }
             }
           } yield {
